+++ src/parsers/blocks/thematic_breaks.jl
@@ -1,3 +1,4 @@
+"""Horizontal rule/thematic break. Build with `Node(ThematicBreak)`."""
 struct ThematicBreak <: AbstractBlock end
 
 accepts_lines(::ThematicBreak) = false
@@ -7,6 +8,8 @@ continue_(::ThematicBreak, ::Parser, ::Node) = 1
 finalize(::ThematicBreak, ::Parser, ::Node) = nothing
 
 can_contain(::ThematicBreak, t) = false
+
+Node(::Type{ThematicBreak}) = Node(ThematicBreak())
 
 function thematic_break(p::Parser, container::Node)
     if !p.indented && occursin(reThematicBreak, rest_from_nonspace(p))